backend/requirements.txt
Flask==3.1.3
dotenv==0.9.9
flask_sqlalchemy==3.1.1
pymysql==1.1.2
cryptography==46.0.7
google-auth==2.38.0